Refactor tests and update DocC
