+++ Sources/Orders/DTOs/OrdersForDeviceDTO.swift
@@ -1,10 +1,3 @@
-//
-//  OrdersForDeviceDTO.swift
-//  PassKit
-//
-//  Created by Francesco Paolo Severino on 30/06/24.
-//
-
 import Vapor
 
 struct OrdersForDeviceDTO: Content {
