+++ Sources/Orders/Middleware/AppleOrderMiddleware.swift
@@ -1,10 +1,3 @@
-//
-//  AppleOrderMiddleware.swift
-//  PassKit
-//
-//  Created by Francesco Paolo Severino on 30/06/24.
-//
-
 import FluentKit
 import Vapor
