+++ Sources/Orders/Models/Concrete Models/OrdersErrorLog.swift
@@ -1,10 +1,3 @@
-//
-//  OrdersErrorLog.swift
-//  PassKit
-//
-//  Created by Francesco Paolo Severino on 30/06/24.
-//
-
 import FluentKit
 import PassKit
