+++ Sources/Orders/Models/OrderModel.swift
@@ -1,10 +1,3 @@
-//
-//  OrderModel.swift
-//  PassKit
-//
-//  Created by Francesco Paolo Severino on 30/06/24.
-//
-
 import FluentKit
 import Foundation
