+++ Sources/Orders/Models/OrdersRegistrationModel.swift
@@ -1,10 +1,3 @@
-//
-//  OrdersRegistrationModel.swift
-//  PassKit
-//
-//  Created by Francesco Paolo Severino on 30/06/24.
-//
-
 import FluentKit
 import PassKit
