+++ Sources/Orders/OrdersService.swift
@@ -1,10 +1,3 @@
-//
-//  OrdersService.swift
-//  PassKit
-//
-//  Created by Francesco Paolo Severino on 01/07/24.
-//
-
 import FluentKit
 import Vapor
