+++ Sources/Passes/Models/PassDataModel.swift
@@ -28,7 +28,7 @@
 
 import FluentKit
 
-/// Represents the `Model` that stores custom app data associated to PassKit passes.
+/// Represents the `Model` that stores custom app data associated to Apple Wallet passes.
 public protocol PassDataModel: Model {
     associatedtype PassType: PassModel
