+++ Tests/PassesTests/PassesTests.swift
@@ -1,4 +1,3 @@
-import FluentKit
 import PassKit
 import Testing
 import XCTVapor
